requirements.txt
biopython==1.83
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
joblib==1.3.2
keras==2.15.0
loguru==0.7.2
numpy==1.26.4
pandas==2.2.1
requests==2.31.0
scikit-learn==1.4.1.post1
tensorflow==2.15.0
